Add new bias functional tests and set groundwork for rediscretization

Closes #650

See merge request jatic/aria/dataeval!594

+++ src/dataeval/utils/shared.py
@@ -95,7 +95,7 @@ def get_classes_counts(labels: NDArray[np.int_]) -> tuple[int, int]:
     M = len(classes)
     if M < 2:
         raise ValueError("Label vector contains less than 2 classes!")
-    N = np.sum(counts).astype(int)
+    N = int(np.sum(counts))
     return M, N
